Split out connection, properties, param classes
