+++ aiopioneer/pioneer_avr.py
@@ -3,62 +3,37 @@
 # pylint: disable=relative-beyond-top-level disable=too-many-lines
 
 import asyncio
-import copy
 import logging
-import re
 import math
 import time
-import traceback
 
 from collections.abc import Callable
 from inspect import isfunction
-from types import MappingProxyType
-from typing import Any
 
 from .param import (
     PARAM_IGNORED_ZONES,
-    PARAM_COMMAND_DELAY,
     PARAM_MAX_SOURCE_ID,
     PARAM_MAX_VOLUME,
     PARAM_MAX_VOLUME_ZONEX,
     PARAM_POWER_ON_VOLUME_BOUNCE,
     PARAM_VOLUME_STEP_ONLY,
     PARAM_IGNORE_VOLUME_CHECK,
-    PARAM_DEBUG_LISTENER,
-    PARAM_DEBUG_RESPONDER,
     PARAM_DEBUG_UPDATER,
     PARAM_DEBUG_COMMAND,
-    PARAM_DEFAULTS,
-    PARAM_DEFAULTS_SYSTEM,
-    PARAM_MODEL_DEFAULTS,
-    PARAM_EXTRA_LISTENING_MODES,
-    PARAM_DISABLED_LISTENING_MODES,
-    PARAM_ENABLED_LISTENING_MODES,
     PARAM_VIDEO_RESOLUTION_MODES,
-    PARAM_ZONE_SOURCES,
+    PARAM_AVAILABLE_LISTENING_MODES,
     PARAM_ENABLED_FUNCTIONS,
     PARAM_DISABLE_AUTO_QUERY,
     PARAM_TUNER_AM_FREQ_STEP,
-    PARAM_QUERY_SOURCES,
-    PARAM_ALL_LISTENING_MODES,
-    PARAM_AVAILABLE_LISTENING_MODES,
 )
 from .commands import PIONEER_COMMANDS
+from .connection import PioneerAVRConnection
 from .exceptions import (
-    PioneerError,
-    AVRUnavailableError,
-    AVRUnknownCommandError,
     AVRResponseTimeoutError,
     AVRCommandError,
-    PioneerErrorFormatText,
 )
-from .util import (
-    merge,
-    sock_set_keepalive,
-    get_backoff_delay,
-    cancel_task,
-    safe_wait_for,
-)
+from .properties import PioneerAVRProperties
+from .util import cancel_task, safe_wait_for
 from .const import (
     Zones,
     TunerBand,
@@ -67,7 +42,6 @@ from .const import (
     DEFAULT_TIMEOUT,
     DEFAULT_SCAN_INTERVAL,
     SOURCE_TUNER,
-    LISTENING_MODES,
     DIMMER_MODES,
     TONE_MODES,
     TONE_DB_VALUES,
@@ -93,12 +67,11 @@ from .const import (
 )
 
 from .parsers import process_raw_response
-from .parsers.response import Response
 
 _LOGGER = logging.getLogger(__name__)
 
 
-class PioneerAVR:
+class PioneerAVR(PioneerAVRConnection, PioneerAVRProperties):
     """Pioneer AVR interface."""
 
     def __init__(
@@ -111,202 +84,44 @@ class PioneerAVR:
     ):
         """Initialise the Pioneer AVR interface."""
         _LOGGER.info("Starting aiopioneer %s", VERSION)
-        _LOGGER.debug(
-            '>> PioneerAVR.__init__(host="%s", port=%s, timeout=%s, params=%s)',
-            host,
-            port,
-            timeout,
-            params,
+        _LOGGER.debug(">> PioneerAVR.__init__()")
+        super().__init__(
+            host=host, port=port, timeout=timeout, scan_interval=scan_interval
         )
-        self._host = host
-        self._port = port
-        self._timeout = timeout
-        self.scan_interval = scan_interval
+        PioneerAVRProperties.__init__(self, params)
 
-        # Public properties
-        self.model = None
+        ## Public properties
         self.software_version = None
         self.mac_addr = None
-
-        self.available = False  # connected to avr
-        self.initial_refresh: list[Zones] = []
         self.zones: list[Zones] = []
-        self.power: dict[Zones, bool] = {}
-        self.volume: dict[Zones, int] = {}
-        self.max_volume: dict[Zones, int] = {}
-        self.mute: dict[Zones, bool] = {}
-        self.source: dict[Zones, str] = {}
-        self.listening_mode = ""
-        self.listening_mode_raw = ""
-        self.media_control_mode: dict[Zones, str] = {}
+        self.initial_refresh: list[Zones] = []
 
-        # FUNC: TONE
-        self.tone: dict[Zones, dict] = {}
-
-        # FUNC: AMP
-        self.amp: dict[str, Any] = {}
-
-        # FUNC: TUNER
-        self.tuner: dict[str, Any] = {}
-
-        # Complex object that holds multiple different props for the CHANNEL/DSP functions
-        self.channel_levels: dict[str, Any] = {}
-        self.dsp: dict[str, Any] = {}
-        self.video: dict[str, Any] = {}
-        self.system: dict[str, Any] = {}
-        self.audio: dict[str, Any] = {}
-
-        # Parameters
-        self._default_params = PARAM_DEFAULTS
-        self._system_params = PARAM_DEFAULTS_SYSTEM
-        self._user_params: dict[str, Any] = None
-        self._params: dict[str, Any] = None
-        self.set_user_params(params)
-
-        # Internal state
-        self._connect_lock = asyncio.Lock()
-        self._disconnect_lock = asyncio.Lock()
+        ## Internal state
         self._update_lock = asyncio.Lock()
-        self._request_lock = asyncio.Lock()
         self._update_event = asyncio.Event()
-        self._response_event = asyncio.Event()
-        self._response_queue: list[Response] = []
-        self._queue_responses = False
-        self._reconnect = True
         self._refresh_zones: list[Zones] = []
-        self._last_updated = None
-        self._last_command = None
-        self._reader = None
-        self._writer = None
-        self._listener_task = None
-        self._responder_task = None
-        self._reconnect_task = None
         self._updater_task = None
         self._command_queue_task = None
         self._command_queue: list[str] = []  # queue of commands to execute
         self._power_zone_1 = None
-        self._source_name_to_id: dict[str, str] = {}
-        self._source_id_to_name: dict[str, str] = {}
         self._zone_callback = {}
 
     def __del__(self):
         _LOGGER.debug(">> PioneerAVR.__del__()")
 
-    @property
-    def query_sources(self) -> bool:
-        """Whether sources have been queried from AVR."""
-        return self._system_params.get(PARAM_QUERY_SOURCES)
+    ## Connection/disconnection
+    async def on_connect(self) -> None:
+        """Start AVR tasks on connection."""
+        await super().on_connect()
+        await self._updater_schedule()
+        await asyncio.sleep(0)  # yield to updater task
 
-    def _set_query_sources(self, value: bool) -> None:
-        self._system_params[PARAM_QUERY_SOURCES] = value
-        self._update_params()
-
-    def get_unique_id(self) -> str:
-        """Get unique identifier for this instance."""
-        return self._host + ":" + str(self._port)
-
-    # Parameter management functions
-    def _update_params(self) -> None:
-        """Set current parameters."""
-        self._params = {}
-        merge(self._params, self._default_params)
-        merge(self._params, self._user_params, force_overwrite=True)
-        if (
-            self._params.get(PARAM_TUNER_AM_FREQ_STEP)
-            and PARAM_TUNER_AM_FREQ_STEP in self._system_params
-        ):
-            ## defer PARAM_TUNER_AM_FREQ_STEP to _user_params if specified
-            del self._system_params[PARAM_TUNER_AM_FREQ_STEP]
-        merge(self._params, self._system_params)
-
-    def set_user_params(self, params: dict[str, Any] = None) -> None:
-        """Set user parameters and update current parameters."""
-        _LOGGER.debug(">> PioneerAVR.set_user_params(%s)", params)
-        self._user_params = copy.deepcopy(params) if params is not None else {}
-        self._update_params()
-        self._update_listening_modes()
-
-    def _set_default_params_model(self) -> None:
-        """Set default parameters based on device model."""
-        model = self.model
-        self._default_params = PARAM_DEFAULTS
-        if model is not None and model != "unknown":
-            for model_regex, model_params in PARAM_MODEL_DEFAULTS.items():
-                if re.search(model_regex, model):
-                    _LOGGER.info(
-                        "applying default parameters for model %s (%s)",
-                        model,
-                        model_regex,
-                    )
-                    merge(self._default_params, model_params, force_overwrite=True)
-        self._update_params()
-
-    def get_param(self, param_name: str) -> Any:
-        """Get the value of the specified parameter."""
-        return self._params.get(param_name)
-
-    def get_params(self) -> dict[str, Any]:
-        """Get a copy of all current parameters."""
-        ## NOTE: can't use MappingProxyTypeType because of mutable dict values
-        return copy.deepcopy(self._params)
-
-    def get_user_params(self) -> dict[str, Any]:
-        """Get a copy of user parameters."""
-        return copy.deepcopy(self._user_params)
-
-    def get_default_params(self) -> dict[str, Any]:
-        """Get a copy of current default parameters."""
-        return copy.deepcopy(self._default_params)
-
-    # Connection/disconnection
-    async def connect(self, reconnect: bool = True) -> None:
-        """Open connection to AVR and start listener thread."""
-        _LOGGER.debug(">> PioneerAVR.connect() started")
-        if self._connect_lock.locked():
-            _LOGGER.warning("AVR connection is already connecting, skipping connect")
-            return
-        if self.available:
-            _LOGGER.warning("AVR is connected, skipping connect")
-            return
-
-        async with self._connect_lock:
-            _LOGGER.debug("opening AVR connection")
-            if self._writer is not None:
-                raise RuntimeError("AVR connection already established")
-
-            # Open new connection
-            reader, writer = await asyncio.wait_for(  # pylint: disable=unused-variable
-                asyncio.open_connection(self._host, self._port), timeout=self._timeout
-            )
-            _LOGGER.info("AVR connection established")
-            self._reader = reader
-            self._writer = writer
-            self.available = True
-            self._reconnect = reconnect
-            self._set_socket_options()
-
-            await self._responder_cancel()
-            await self._listener_schedule()
-            await asyncio.sleep(0)  # yield to listener task
-            await self._updater_schedule()
-            await asyncio.sleep(0)  # yield to updater task
-
-        _LOGGER.debug(">> PioneerAVR.connect() completed")
-
-    def _set_socket_options(self) -> None:
-        """Set socket keepalive options."""
-        sock_set_keepalive(
-            self._writer.get_extra_info("socket"),
-            after_idle_sec=int(self._timeout),
-            interval_sec=int(self._timeout),
-            max_fails=3,
-        )
-
-    async def set_timeout(self, timeout: float) -> None:
-        """Set timeout and update socket keepalive options."""
-        _LOGGER.debug(">> PioneerAVR.set_timeout(%f)", timeout)
-        self._timeout = timeout
-        self._set_socket_options()
+    async def on_disconnect(self) -> None:
+        """Stop AVR"""
+        self._call_zone_callbacks()
+        await self._command_queue_cancel()
+        await self._updater_cancel()
+        await super().on_disconnect()
 
     async def set_scan_interval(self, scan_interval: int) -> None:
         """Set scan interval and restart updater."""
@@ -316,425 +131,7 @@ class PioneerAVR:
             self.scan_interval = scan_interval
             await self._updater_schedule()
 
-    async def disconnect(self) -> None:
-        """Shutdown and close connection to AVR."""
-        _LOGGER.debug(">> PioneerAVR.disconnect() started")
-
-        if self._disconnect_lock.locked():
-            _LOGGER.warning(
-                "AVR connection is already disconnecting, skipping disconnect"
-            )
-            return
-        if not self.available:
-            _LOGGER.warning("AVR not connected, skipping disconnect")
-            return
-
-        async with self._disconnect_lock:
-            _LOGGER.debug("disconnecting AVR connection")
-            self.available = False
-            self._call_zone_callbacks()
-
-            await self._command_queue_cancel()
-            await self._updater_cancel()
-            await self._responder_cancel()
-            await self._listener_cancel()
-
-            writer = self._writer
-            if writer:
-                # Close AVR connection
-                _LOGGER.debug("closing AVR connection")
-                self._writer.close()
-                try:
-                    await self._writer.wait_closed()
-                except Exception as exc:  # pylint: disable=broad-except
-                    _LOGGER.debug("ignoring disconnect exception: %s", str(exc))
-            self._reader = None
-            self._writer = None
-            _LOGGER.info("AVR connection closed")
-
-            await self._reconnect_schedule()
-
-        _LOGGER.debug(">> PioneerAVR.disconnect() completed")
-
-    async def shutdown(self) -> None:
-        """Shutdown the client."""
-        _LOGGER.debug(">> PioneerAVR.shutdown()")
-        self._reconnect = False
-        await self._reconnect_cancel()
-        await self.disconnect()
-        await asyncio.sleep(0)  # yield to pending shutdown tasks
-
-    async def reconnect(self) -> None:
-        """Reconnect to an AVR."""
-        _LOGGER.debug(">> PioneerAVR.reconnect() started")
-        retry = 0
-        try:
-            while not self.available:
-                delay = get_backoff_delay(retry)
-                _LOGGER.debug("waiting %ds before retrying connection", delay)
-                await asyncio.sleep(delay)
-
-                retry += 1
-                try:
-                    await self.connect()
-                    if self.available:
-                        break
-                except asyncio.CancelledError:  # pylint: disable=try-except-raise
-                    # pass through to outer except
-                    raise
-                except Exception as exc:  # pylint: disable=broad-except
-                    _LOGGER.debug(
-                        "could not reconnect to AVR: %s: %s", type(exc).__name__, exc
-                    )
-                    # fall through to reconnect outside try block
-
-                if self.available:
-                    await self.disconnect()
-        except asyncio.CancelledError:
-            _LOGGER.debug(">> PioneerAVR.reconnect() cancelled")
-
-        _LOGGER.debug(">> PioneerAVR.reconnect() completed")
-
-    async def _reconnect_schedule(self) -> None:
-        """Schedule reconnection to the AVR."""
-        if self._reconnect:
-            _LOGGER.debug(">> PioneerAVR._reconnect_schedule()")
-            reconnect_task = self._reconnect_task
-            if reconnect_task:
-                await asyncio.sleep(0)  # yield to reconnect task if running
-                if reconnect_task.done():
-                    reconnect_task = None  # trigger new task creation
-            if reconnect_task is None:
-                _LOGGER.info("reconnecting to AVR")
-                reconnect_task = asyncio.create_task(self.reconnect())
-                self._reconnect_task = reconnect_task
-            else:
-                _LOGGER.error("AVR listener reconnection already running")
-
-    async def _reconnect_cancel(self) -> None:
-        """Cancel any active reconnect task."""
-        await cancel_task(self._reconnect_task, "reconnect")
-        self._reconnect_task = None
-
-    async def _connection_listener(self) -> None:
-        """AVR connection listener. Parse responses and update state."""
-        if self._params[PARAM_DEBUG_LISTENER]:
-            _LOGGER.debug(">> PioneerAVR._connection_listener() started")
-        running = True
-        while self.available:
-            debug_listener = self._params[PARAM_DEBUG_LISTENER]
-            action = " listening for responses"
-            try:
-                response = await self._read_response()
-                if response is None:
-                    # Connection closed or exception, exit task
-                    break
-
-                # Check for empty response
-                self._last_updated = time.time()  # include empty responses
-                if not response:
-                    # Skip processing empty responses (keepalives)
-                    if debug_listener:
-                        _LOGGER.debug("ignoring empty response")
-                    continue
-                if debug_listener:
-                    _LOGGER.debug("received AVR response: %s", response)
-
-                # Parse response, update cached properties
-                action = " parsing response " + response
-                parse_result = self._parse_response(response)
-                updated_zones = parse_result.get("updated_zones")
-
-                ## Queue raw response and signal response handler
-                if self._queue_responses:
-                    self._response_queue.append(response)
-                    self._response_event.set()
-                    ## Do not yield, process all responses first
-
-                ## Detect Zone 1 power on for volume workaround
-                action = ""
-                power_on_volume_bounce = self._params[PARAM_POWER_ON_VOLUME_BOUNCE]
-                if power_on_volume_bounce and self._power_zone_1 is not None:
-                    if not self._power_zone_1 and self.power.get(Zones.Z1):
-                        ## Zone 1 powered on, schedule bounce task
-                        _LOGGER.info("scheduling main zone volume workaround")
-                        self.queue_command(
-                            "volume_up", skip_if_queued=False, insert_at=0
-                        )
-                        self.queue_command(
-                            "volume_down", skip_if_queued=False, insert_at=1
-                        )
-                self._power_zone_1 = self.power.get(Zones.Z1)  # cache value
-
-                # Implement a command queue so that we can queue commands if we
-                # need to update attributes that only get updated when we
-                # request them to change.
-                self.command_queue_schedule()
-
-                # NOTE: to avoid deadlocks, do not run any operations that
-                # depend on further responses (returned by the listener) within
-                # the listener loop.
-
-                if updated_zones:
-                    action = f" while calling zone callbacks for {updated_zones}"
-                    # Call zone callbacks for updated zones
-                    self._call_zone_callbacks(updated_zones)
-                    # NOTE: updating zone 1 does not reset its scan interval -
-                    # scan interval is set to a regular timer
-
-            except asyncio.CancelledError:
-                if debug_listener:
-                    _LOGGER.debug(">> PioneerAVR._connection_listener() cancelled")
-                running = False
-                break
-            except Exception as exc:  # pylint: disable=broad-except
-                _LOGGER.error("listener exception%s: %s", action, str(exc))
-                _LOGGER.error(traceback.format_exc())
-                # continue listening on exception
-
-        if running and self.available:
-            # Trigger disconnection if not already disconnected
-            await self.disconnect()
-
-        _LOGGER.debug(">> PioneerAVR._connection_listener() completed")
-
-    async def _listener_schedule(self) -> None:
-        """Schedule the listener task."""
-        if self._params[PARAM_DEBUG_LISTENER]:
-            _LOGGER.debug(">> PioneerAVR._listener_schedule()")
-        await self._listener_cancel()
-        self._listener_task = asyncio.create_task(self._connection_listener())
-
-    async def _listener_cancel(self) -> None:
-        """Cancel the listener task."""
-        debug_listener = self._params[PARAM_DEBUG_LISTENER]
-        await cancel_task(self._listener_task, "listener", debug=debug_listener)
-        self._listener_task = None
-
-    # Reader co-routine
-    async def _reader_readuntil(self) -> bytes | None:
-        """Read from reader with cancel detection."""
-        try:
-            return await self._reader.readuntil(b"\n")
-        except asyncio.CancelledError:
-            if self._params[PARAM_DEBUG_RESPONDER]:
-                _LOGGER.debug("reader: readuntil() was cancelled")
-            return None
-
-    # Read responses from AVR
-    async def _read_response(self, timeout: float = None) -> str:
-        """Wait for a response from AVR and return to all readers."""
-        debug_responder = self._params[PARAM_DEBUG_RESPONDER]
-        if debug_responder:
-            _LOGGER.debug(">> PioneerAVR._read_response(timeout=%s)", timeout)
-
-        # Schedule responder task if not already created
-        responder_task = self._responder_task
-        if responder_task:
-            if responder_task.done():
-                responder_task = None  # trigger new task creation
-        if responder_task is None:
-            responder_task = asyncio.create_task(self._reader_readuntil())
-            # responder_task = asyncio.create_task(self._reader.readuntil(b"\n"))
-            self._responder_task = responder_task
-            if debug_responder:
-                _LOGGER.debug("created responder task %s", responder_task.get_name())
-        else:
-            # Wait on existing responder task
-            if debug_responder:
-                _LOGGER.debug("using responder task %s", responder_task.get_name())
-
-        # Wait for result and process
-        task_name = asyncio.current_task().get_name()
-        try:
-            if timeout:
-                if debug_responder:
-                    _LOGGER.debug(
-                        "%s: waiting for data (timeout=%s)", task_name, timeout
-                    )
-                done, pending = await asyncio.wait(  # pylint: disable=unused-variable
-                    [responder_task], timeout=timeout
-                )
-                if done:
-                    raw_response = responder_task.result()
-                else:
-                    _LOGGER.debug("%s: timed out waiting for data", task_name)
-                    return None
-            else:
-                if debug_responder:
-                    _LOGGER.debug("%s: waiting for data", task_name)
-                raw_response = await responder_task
-        except (EOFError, TimeoutError):
-            # Connection closed
-            if debug_responder:
-                _LOGGER.debug("%s: connection closed", task_name)
-            return None
-        except Exception as exc:  # pylint: disable=broad-except
-            _LOGGER.error("%s: exception: %s", task_name, str(exc))
-            return None
-        if raw_response is None:  # task cancelled
-            return None
-        response = raw_response.decode().strip()
-        if debug_responder:
-            _LOGGER.debug("%s: received response: %s", task_name, response)
-        return response
-
-    async def _responder_cancel(self) -> None:
-        """Cancel any active responder task."""
-        debug_responder = self._params[PARAM_DEBUG_RESPONDER]
-        await cancel_task(self._responder_task, "responder", debug=debug_responder)
-        self._responder_task = None
-
-    # Send commands and requests to AVR
-    async def send_raw_command(self, command: str, rate_limit: bool = True) -> None:
-        """Send a raw command to the AVR."""
-        debug_command = self._params[PARAM_DEBUG_COMMAND]
-        if debug_command:
-            _LOGGER.debug(
-                '>> PioneerAVR.send_raw_command("%s", rate_limit=%s)',
-                command,
-                rate_limit,
-            )
-        if not self.available:
-            raise AVRUnavailableError
-
-        if rate_limit:
-            # Rate limit commands
-            command_delay = self._params[PARAM_COMMAND_DELAY]
-            since_command = command_delay + 0.1
-            if self._last_command:
-                since_command = time.time() - self._last_command
-            if since_command < command_delay:
-                delay = command_delay - since_command
-                if debug_command:
-                    _LOGGER.debug("delaying command for %.3f s", delay)
-                await asyncio.sleep(command_delay - since_command)
-        _LOGGER.debug("sending command: %s", command)
-        self._writer.write(command.encode("ASCII") + b"\r")
-        await self._writer.drain()
-        self._last_command = time.time()
-
-    async def _wait_for_response(self, command: str, response_prefix: str) -> str:
-        """Wait for a response to a request."""
-        debug_command = self._params[PARAM_DEBUG_COMMAND]
-
-        while True:
-            await self._response_event.wait()
-            self._response_event.clear()
-            for response in self._response_queue:
-                if response.startswith(response_prefix):
-                    if debug_command:
-                        _LOGGER.debug(
-                            "AVR command %s returned response: %s", command, response
-                        )
-                    return response
-                if response.startswith("E"):
-                    raise AVRCommandError(response)
-            self._response_queue = []
-
-    async def send_raw_request(
-        self,
-        command: str,
-        response_prefix: str,
-        rate_limit: bool = True,
-    ) -> str:
-        """Send a raw command to the AVR and return the response."""
-        debug_command = self._params[PARAM_DEBUG_COMMAND]
-        if debug_command:
-            _LOGGER.debug(
-                '>> PioneerAVR.send_raw_request("%s", %s, rate_limit=%s)',
-                command,
-                response_prefix,
-                rate_limit,
-            )
-        async with self._request_lock:  ## Only send one request at a time
-            self._response_queue = []
-            self._queue_responses = (
-                True  ## Start queueing responses before sending command
-            )
-            self._response_event.clear()
-            await self.send_raw_command(command, rate_limit=rate_limit)
-            try:
-                # response = await asyncio.wait_for(
-                response = await safe_wait_for(
-                    self._wait_for_response(command, response_prefix),
-                    timeout=self._timeout,
-                )
-                await asyncio.sleep(0)  # yield to listener task
-            except TimeoutError as exc:  # response timer expired
-                raise AVRResponseTimeoutError from exc
-
-            self._queue_responses = False
-            self._response_queue = []
-            return response
-
-    async def send_command(
-        self,
-        command: str,
-        zone: Zones = Zones.Z1,
-        prefix: str = "",
-        suffix: str = "",
-        ignore_error: bool | None = None,
-        rate_limit: bool = True,
-    ) -> str | bool | None:
-        """Send a command or request to the device."""
-        # pylint: disable=unidiomatic-typecheck disable=logging-not-lazy
-        debug_command = self._params[PARAM_DEBUG_COMMAND]
-        if debug_command:
-            _LOGGER.debug(
-                '>> PioneerAVR.send_command("%s", zone="%s", prefix="%s", '
-                + "ignore_error=%s, rate_limit=%s, suffix=%s)",
-                command,
-                zone,
-                prefix,
-                ignore_error,
-                rate_limit,
-                suffix,
-            )
-
-        async def _send_command():
-            raw_command = PIONEER_COMMANDS.get(command, {}).get(zone)
-            if type(raw_command) is list:
-                if len(raw_command) == 2:
-                    # Handle command as request
-                    expected_response = raw_command[1]
-                    raw_command = raw_command[0]
-                    response = await self.send_raw_request(
-                        prefix + raw_command + suffix,
-                        response_prefix=expected_response,
-                        rate_limit=rate_limit,
-                    )
-                    if debug_command:
-                        _LOGGER.debug("send_command received response: %s", response)
-                    return response
-                raise AVRUnknownCommandError
-            elif type(raw_command) is str:
-                await self.send_raw_command(
-                    prefix + raw_command + suffix, rate_limit=rate_limit
-                )
-                return True
-            raise AVRUnknownCommandError
-
-        if ignore_error is None:
-            ## Do not handle exceptions
-            return await _send_command()
-
-        # pylint: disable=broad-exception-caught
-        try:
-            return await _send_command()
-        except (PioneerError, Exception) as exc:
-            translation_key = getattr(exc, "translation_key", "unknown_exception")
-            err = PioneerErrorFormatText.get(translation_key, "unknown_exception")
-            err_txt = err.format(command=command, zone=str(zone), exc=str(exc))
-            rc = False if isinstance(exc, AVRCommandError) else None
-
-        if ignore_error:
-            _LOGGER.debug(err_txt)
-        else:
-            _LOGGER.error(err_txt)
-        return rc
-
-    # Initialisation functions
+    ## Initialisation functions
     async def query_zones(self, force_update: bool = False) -> None:
         """Query zones on Pioneer AVR by querying power status."""
         _LOGGER.info("querying available zones on AVR")
@@ -778,12 +175,6 @@ class PioneerAVR:
                 removed_zones = True
         await self.query_zones(force_update=removed_zones)
 
-    def set_source_dict(self, sources: dict[str, str]) -> None:
-        """Manually set source id<->name translation tables."""
-        self._set_query_sources(False)
-        self._source_name_to_id = copy.deepcopy(sources)
-        self._source_id_to_name = {v: k for k, v in sources.items()}
-
     async def build_source_dict(self) -> None:
         """Generate source id<->name translation tables."""
         timeouts = 0
@@ -812,47 +203,6 @@ class PioneerAVR:
         if not self._source_name_to_id:
             _LOGGER.warning("no input sources found on AVR")
 
-    def get_source_list(self, zone: Zones = Zones.Z1) -> list[str]:
-        """Return list of available input sources."""
-        source_ids = self._params.get(PARAM_ZONE_SOURCES[zone], [])
-        return list(
-            self._source_name_to_id.keys()
-            if not source_ids
-            else [
-                self._source_id_to_name[s]
-                for s in source_ids
-                if s in self._source_id_to_name
-            ]
-        )
-
-    def get_source_dict(self, zone: Zones = None) -> dict[str, str]:
-        """Return source id<->name translation tables."""
-        if zone is None:
-            return MappingProxyType(self._source_name_to_id)
-        source_ids = self._params.get(PARAM_ZONE_SOURCES[zone], [])
-        return (
-            self._source_name_to_id
-            if not source_ids
-            else {k: v for k, v in self._source_name_to_id.items() if v in source_ids}
-        )
-
-    def get_source_name(self, source_id: str) -> str:
-        """Return name for given source ID."""
-        return (
-            self._source_id_to_name.get(source_id, source_id)
-            if self._source_name_to_id
-            else source_id
-        )
-
-    def clear_source_id(self, source_id: str) -> None:
-        """Clear name mapping for given source ID."""
-        source_name = None
-        if source_id in self._source_id_to_name:
-            source_name = self._source_id_to_name[source_id]
-            self._source_id_to_name.pop(source_id)
-        if source_name in self._source_name_to_id:
-            self._source_name_to_id.pop(source_name)
-
     def get_listening_modes(self) -> dict[str, str] | None:
         """Return dict of valid listening modes and names for Zone 1."""
         multichannel = self.audio.get("input_multichannel")
@@ -864,70 +214,6 @@ class PioneerAVR:
             ):
                 zone_listening_modes |= {mode_id: mode_details[0]}
         return zone_listening_modes
-
-    def _update_listening_modes(self) -> None:
-        """Update list of valid listening modes for AVR."""
-        all_listening_modes = LISTENING_MODES | self._params.get(
-            PARAM_EXTRA_LISTENING_MODES, {}
-        )
-        disabled_listening_modes = self._params.get(PARAM_DISABLED_LISTENING_MODES)
-        enabled_listening_modes = self._params.get(PARAM_ENABLED_LISTENING_MODES)
-        available_listening_modes = {}
-        available_listening_mode_names = []
-
-        for mode_id, mode_details in all_listening_modes.items():
-            if mode_id in disabled_listening_modes or (
-                enabled_listening_modes and mode_id not in enabled_listening_modes
-            ):
-                pass
-            elif mode_details[0] in available_listening_mode_names:
-                _LOGGER.error(
-                    "ignored duplicate listening mode name: %s", mode_details[0]
-                )
-            else:
-                available_listening_modes |= {mode_id: mode_details}
-                available_listening_mode_names.append(mode_details[0])
-
-        _LOGGER.info("determining available listening modes")
-        self._system_params[PARAM_ALL_LISTENING_MODES] = all_listening_modes
-        self._system_params[PARAM_AVAILABLE_LISTENING_MODES] = available_listening_modes
-        self._update_params()
-
-    def get_ipod_control_commands(self) -> list[str]:
-        """Return a list of all valid iPod control modes."""
-        return list(
-            [
-                k.replace("operation_ipod_", "")
-                for k in PIONEER_COMMANDS
-                if k.startswith("operation_ipod")
-            ]
-        )
-
-    def get_tuner_control_commands(self) -> list[str]:
-        """Return a list of all valid tuner control commands."""
-        return list(
-            [
-                k.replace("operation_tuner_", "")
-                for k in PIONEER_COMMANDS
-                if k.startswith("operation_tuner")
-            ]
-        )
-
-    def get_supported_media_controls(self, zone: Zones) -> list[str] | None:
-        """Return a list of all valid media control actions for a given zone.
-        If the provided zone source is not currently compatible with media controls,
-        null will be returned."""
-        if self.media_control_mode.get(zone) is not None:
-            return list(
-                [
-                    k
-                    for k in MEDIA_CONTROL_COMMANDS.get(
-                        self.media_control_mode.get(zone)
-                    ).keys()
-                ]
-            )
-        else:
-            return None
 
     async def query_device_info(self) -> None:
         """Query device information from Pioneer AVR."""
@@ -949,8 +235,7 @@ class PioneerAVR:
         # It will report software version only if Zone 1 is powered on.
         # It is unknown how iControlAV5 determines this on a routed network.
 
-    ## Callback functions
-
+    ## Client callback functions
     def set_zone_callback(
         self, zone: Zones, callback: Callable[..., None] | None = None
     ) -> None:
@@ -975,8 +260,8 @@ class PioneerAVR:
                 if callback:
                     callback()
 
-    # Update functions
-    def _parse_response(self, response_raw: str) -> dict[str, Any]:
+    ## Response handling callbacks
+    def parse_response(self, response_raw: str) -> set:
         """Parse response and update cached parameters."""
         updated_zones = set()
 
@@ -1083,10 +368,38 @@ class PioneerAVR:
                 ):
                     self.queue_command("_delayed_query_av_information")
 
-        result = {"updated_zones": updated_zones}
+        return updated_zones
 
-        return result
+    def handle_response(self, updated_zones: set) -> None:
+        """Handle responses."""
+        ## TODO: move workarounds in parse_response to here
 
+        ## Detect Zone 1 power on for volume workaround
+        power_on_volume_bounce = self._params[PARAM_POWER_ON_VOLUME_BOUNCE]
+        if power_on_volume_bounce and self._power_zone_1 is not None:
+            if not self._power_zone_1 and self.power.get(Zones.Z1):
+                ## Zone 1 powered on, schedule bounce task
+                _LOGGER.info("scheduling main zone volume workaround")
+                self.queue_command("volume_up", skip_if_queued=False, insert_at=0)
+                self.queue_command("volume_down", skip_if_queued=False, insert_at=1)
+        self._power_zone_1 = self.power.get(Zones.Z1)  # cache value
+
+        # Implement a command queue so that we can queue commands if we
+        # need to update attributes that only get updated when we
+        # request them to change.
+        self.command_queue_schedule()
+
+        # NOTE: to avoid deadlocks, do not run any operations that
+        # depend on further responses (returned by the listener) within
+        # the listener loop.
+
+        if updated_zones:
+            # Call zone callbacks for updated zones
+            self._call_zone_callbacks(updated_zones)
+            # NOTE: updating zone 1 does not reset its scan interval -
+            # scan interval is set to a regular timer
+
+    ## AVR Updater
     async def _updater(self) -> None:
         """Perform update every scan_interval."""
         debug_updater = self._params[PARAM_DEBUG_UPDATER]
@@ -1196,13 +509,13 @@ class PioneerAVR:
 
         _rc = True
         async with self._update_lock:
-            # Update only if scan_interval has passed
+            ## Update only if scan_interval has passed
             now = time.time()
             refresh_zones = self._refresh_zones
             since_refreshed = self.scan_interval + 1
             since_refreshed_str = "never"
-            if self._last_updated:
-                since_refreshed = now - self._last_updated
+            if self.last_updated:
+                since_refreshed = now - self.last_updated
                 since_refreshed_str = f"{since_refreshed:.3f}s ago"
             if not self.scan_interval or since_refreshed > self.scan_interval:
                 refresh_zones = self.zones
@@ -1213,7 +526,7 @@ class PioneerAVR:
                     refresh_zones,
                     since_refreshed_str,
                 )
-                self._last_updated = now
+                self.last_updated = now
                 try:
                     for zone in refresh_zones:
                         await self._refresh_zone(zone)
@@ -1272,8 +585,7 @@ class PioneerAVR:
         else:
             _LOGGER.error("AVR updater task not running")
 
-    # State change functions
-
+    ## State change functions
     def _get_parameter_key_from_value(
         self, val: str, parameters: dict, loose_match: bool = False
     ) -> str:
@@ -1326,6 +638,7 @@ class PioneerAVR:
         zone = self._check_zone(zone)
         await self.send_command("volume_down", zone)
 
+    ## Command queue
     async def _execute_command_queue(self) -> None:
         """Execute commands from a queue."""
         debug_command = self._params[PARAM_DEBUG_COMMAND]
@@ -1441,6 +754,7 @@ class PioneerAVR:
         else:
             self._command_queue.append(command)
 
+    ## Additional AVR methods
     async def set_volume_level(
         self, target_volume: int, zone: Zones = Zones.Z1
     ) -> bool:
@@ -1683,7 +997,7 @@ class PioneerAVR:
             return
 
         self._system_params[PARAM_TUNER_AM_FREQ_STEP] = new_freq - current_freq
-        self._update_params()
+        self.update_params()
         await self.send_command("decrease_tuner_frequency", ignore_error=True)
 
     async def _step_tuner_frequency(self, band: str, frequency: float) -> None:
